ALSA: compress: fix an integer overflow check
Change-Id: Iba5acd089cae2d5320ddc82789e5e5b69a4cfdc1

I previously added an integer overflow check here but looking at it now,
it's still buggy. The bug happens in snd_compr_allocate_buffer().
We multiply ".fragments" and ".fragment_size" and that doesn't overflow
but then we save it in an unsigned int so it truncates the high bits
away and we allocate a smaller than expected size.

Fixes: b35cc8225845 ('ALSA: compress_core: integer overflow in snd_compr_allocate_buffer()')
Change-Id: I2a7dccf9d5b6d7bd02e841c1a6fa754bf4978cd8
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>