FPII-2523: Elevation of privilege vulnerability in kernel performance subsystem CVE-2015-8963 A-30952077
perf: Fix race in swevent hash
There's a race on CPU unplug where we free the swevent hash array
while it can still have events on. This will result in a
use-after-free which is BAD.

Simply do not free the hash array on unplug. This leaves the thing
around and no use-after-free takes place.

When the last swevent dies, we do a for_each_possible_cpu() iteration
anyway to clean these up, at which time we'll free it, so no leakage
will occur.
Change-Id: I5032ac3b18ce45e0256c8f9af9c2fa6e9cd887a2