FPII-2591: Elevation of privilege vulnerability in kernel (device specific) CVE-2015-8966 A-31435731
An elevation of privilege vulnerability in the kernel could enable a local malicious
application to execute arbitrary code within the context of the kernel. This issue is
rated as Critical due to the possibility of a local permanent device compromise,
which may require reflashing the operating system to repair the device.

Additional technical details:

Bug Details
 A-31435731
In the sys_oabi_fcntl64 function, it is possible for the KERNEL_DS to be set and not be
set back to USER_DS, leading to arbitrary memory access in the kernel.

The fix is designed to make sure the USER_DS is set back to prevent the possible arbitrary memory access in the kernel.

Change-Id: I46401e2369e7780e71b0adba4e025e8892fc21bd