wlan: Fix reorder packets PN replay check algorithm
Currently for reorder packets PN replay check of the out
of order missing packet is done with the last packet received
in the reorder window slot. The PN number of the out of order
missing packet will be less than the last received packet in
the window and packet drop happens because of replay check.

In this change fix the logic by doing the replay check of the
out of order missing packet with the previous sequence packet
that of out of order packet.

Issue: SEC-2319
Change-Id: Ib83b825d4b9b292e125c98d819fdfa6eb160e389
CRs-Fixed: 2145435
(cherry picked from commit 6eb0eaeba87f4f12eeaa3f072a11c46e05a5c648)