Revert "FP2: dont hardreset PMIC"
This reverts commit e8ab366049f92612f4926ea422a1ce313625e305.

The kernel boot reason does not get updated when hard resets occur.
This causes Hiccup crash reports with outdated boot reasons to be sent.
Reverting e8ab36604 makes the hard resets being reported as
"keyboard power on" again, as it was the case in Android 6.

Issue: HIC-244
Issue: FP2N-231
Change-Id: I7a4d232cca292b3215944abd4cb4bfb794154396