Changes for initial User build bootup [6/7]
This takes the current fairphone_defconfig (used for userdebug builds)
as reference for fairphone-perf_defconfig and disabled configurations
that are not working yet on the user build.

Re-enabling disabled configurations and tweaks for user build (removal
of debug options, optimizing for performance) are to be done in
follow-up commits.

Issue: FP2P-150
Change-Id: Ifdf6c04f11122b41a448d41fd1e94d83edb44888