Fix conflicting implementations after reverts for fixing VPN
The FP2 kernel doesn't implement UID routing properly, which causes a
few CTS test cases to fail. Revert both kernel and system/netd back to
implementations used in Android 7 that still work reliably enough with
the old kernel. This is a hack, but as clean as it gets for now on the
code base we have.

Issue: FP2P-307
Change-Id: I2a8a178c7d14dbef2314d2506ab1483efb2c7bfa